requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
django-scheduler==0.9.6
dlib==19.23.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.1
Flask-APScheduler==1.12.3
Flask-Cors==3.0.10
gunicorn==20.1.0
icalendar==4.0.9
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
yapf==0.32.0
zipp==3.8.0
